scripts/build_index.py
"""
벡터 DB 인덱싱 스크립트

이 스크립트는 data/processed/chunks.json을 읽어서
메타데이터를 포함한 벡터 DB를 구축합니다.

사용법:
    python scripts/build_index.py
"""
import os
import sys

# 프로젝트 루트를 Python 경로에 추가
sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.abspath(__file__))))

import json
import faiss
import numpy as np
from ingest.embed import embed_chunks
from config import FAISS_INDEX_PATH

def build_vector_index():
    """
    chunks.json을 읽어서 FAISS 인덱스를 구축합니다.
    """
    # 1. 청크 데이터 로드
    chunks_path = "data/processed/chunks.json"
    if not os.path.exists(chunks_path):
        print(f"❌ Error: {chunks_path} 파일이 없습니다.")
        print("   먼저 'python scripts/chunk_papers.py'를 실행하세요.")
        return
    
    print(f"📖 Loading chunks from {chunks_path}...")
    with open(chunks_path, "r", encoding="utf-8") as f:
        chunks = json.load(f)
    
    print(f"   총 {len(chunks)}개의 청크를 로드했습니다.")
    
    # 2. 텍스트만 추출하여 임베딩 생성
    print("\n🔢 Creating embeddings...")
    texts = [chunk["text"] for chunk in chunks]
    vectors = embed_chunks(texts)
    
    print(f"   임베딩 차원: {vectors.shape}")
    
    # 3. FAISS 인덱스 구축
    print("\n🗄️  Building FAISS index...")
    dim = vectors.shape[1]
    index = faiss.IndexFlatL2(dim)
    
    # 벡터를 float32로 변환 (FAISS 요구사항)
    vectors = vectors.astype('float32')
    index.add(vectors)
    
    # 4. 인덱스 저장
    index_dir = os.path.dirname(FAISS_INDEX_PATH)
    os.makedirs(index_dir, exist_ok=True)
    
    faiss.write_index(index, FAISS_INDEX_PATH)
    print(f"   ✅ 인덱스 저장 완료: {FAISS_INDEX_PATH}")
    
    # 5. 청크 메타데이터도 함께 저장 (검색 시 사용)
    metadata_path = "data/processed/chunks_metadata.json"
    with open(metadata_path, "w", encoding="utf-8") as f:
        json.dump(chunks, f, ensure_ascii=False, indent=2)
    print(f"   ✅ 메타데이터 저장 완료: {metadata_path}")
    
    print(f"\n✨ 총 {len(chunks)}개의 청크가 인덱싱되었습니다!")
    print(f"   이제 'python main.py'로 RAG를 실행할 수 있습니다.")

if __name__ == "__main__":
    build_vector_index()
