requirements.txt
langchain
langchain-text-splitters
sentence-transformers
faiss-cpu
transformers
torch
datasets
requests